requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.20.0
Pillow>=9.0.0
matplotlib>=3.5.0
opencv-python>=4.5.0
tqdm>=4.60.0
tabulate>=0.8.9
requests>=2.25.0

# YOLO dependencies
ultralytics>=8.0.0

# Supervision dependencies
supervision>=0.13.0

# CLIP dependencies
ftfy
regex
git+https://github.com/openai/CLIP.git

# DINO and GLIP dependencies
transformers>=4.26.0

# Optional dependencies for visualization
seaborn>=0.11.0